seven_segment_display_DEMO/requirements.txt
pyserial==3.5
PyVirtualSerialPorts==1.0.1
seven_segment_display==1.0.0